// frontend-study-buddy/src/api/types.ts
export type UUID = string;

export type TipoUsuario = "UNIVERSITARIO" | "LOCADOR" | "ESTUDANTE";

export type Genero =
  | "MASCULINO"
  | "FEMININO"
  | "OUTRO"
  | "PREFIRO_NAO_INFORMAR";

export type TipoMoradia = "APARTAMENTO" | "CASA" | "REPUBLICA" | "QUARTO";

export type TipoAnuncio = "IMOVEL_COMPLETO" | "VAGA_COMPARTILHADA";

export type StatusAnuncio = "ATIVO" | "PAUSADO" | "ENCERRADO";
// ...
export interface FiltroBuscaAnuncio {
  valorMin?: number;
  valorMax?: number;
  bairro?: string;
  tipoMoradia?: TipoMoradia;
  tipoAnuncio?: TipoAnuncio;
  mobiliado?: boolean;
  aceitaAnimais?: boolean;
  quantidadeVagas?: number;
  page?: number;
  size?: number;
}

export interface BuscaAnuncioResponse {
  id: UUID;
  titulo: string;
  descricao: string;
  valorMensal: number;
  tipoAnuncio: TipoAnuncio;
  status: StatusAnuncio;
  dataPublicacao: string;
  moradiaId: UUID;
  tipoMoradia: TipoMoradia;
  bairro: string;
  enderecoResumo: string;
  mobiliado: boolean;
  aceitaAnimais: boolean;
  quantidadeVagas: number;
  nomeAnunciante: string;
}

export interface PaginaResponse<T> {
  conteudo: T[];
  paginaAtual: number;
  totalPaginas: number;
  totalElementos: number;
  tamanhoPagina: number;
}